requirements.txt
Flask==1.0.2
matplotlib==3.1.1
numpy==1.14.5
psycopg2-binary==2.8
jellyfish==0.6.1
tqdm==4.36.1
torch==1.3.0
python-dotenv==0.10.3
scikit_learn==0.21.3
requests==2.22.0
uWSGI==2.0.18 
flask-cors==3.0.8